🏗️ refactor [#10.3.8]: 테스트 구조 2차 개선 - Unit/Integration 분리, 동시성 안전성 강화
🔍 변경 사항:
- **테스트 디렉토리 재구성**:
    - [tests/unit/test_sync_map_manager.py]: SyncMapManager CRUD + Concurrency (New)
    - [tests/unit/test_obsidian_sync_unit.py]: 해시, FileWatcher, 모델 검증 (New)
    - [tests/integration/test_obsidian_sync.py]: End-to-End 시나리오만 유지 (Modified)
    - [tests/integration/test_sync_map_manager.py]: 삭제 (백업 완료)
- **동시성 테스트 안전성 강화**:
    - `timeout=10` 추가 (교착 상태 방지)
    - `return_when=FIRST_EXCEPTION` 추가 (첫 예외 즉시 감지)
- **Assertion 명확화**: 미완료 개수 표시, 예외 메시지 포맷 개선

📁 파일:
- tests/unit/test_sync_map_manager.py (New)
- tests/unit/test_obsidian_sync_unit.py (New)
- tests/integration/test_obsidian_sync.py (Modified)
- tests/integration/test_sync_map_manager.py (Deleted)

📌 Related
- Issue [#76]
- @ai-bot-review (https://github.com/jjaayy2222/flownote-mvp/pull/132#pullrequestreview-3550386194)
